:sparkles: Change git mob to work globally

+++ test-helpers/index.js
@@ -4,8 +4,8 @@ const { stripIndent } = require('common-tags');
 const eol = require('eol');
 
 function retainLocalAuthor() {
-  const localName = exec('git config --local user.name').stdout.trim();
-  const localEmail = exec('git config --local user.email').stdout.trim();
+  const localName = exec('git config user.name').stdout.trim();
+  const localEmail = exec('git config user.email').stdout.trim();
   return function () {
     if (localEmail && localName) {
       addAuthor(localName, localEmail);
@@ -26,29 +26,35 @@ function removeAuthor() {
 }
 
 function addCoAuthor(name, email) {
-  exec(`git config --add git-mob.co-author "${name} <${email}>"`);
+  exec(`git config --global --add git-mob.co-author "${name} <${email}>"`);
 }
 
 function removeCoAuthors() {
-  exec('git config --unset-all git-mob.co-author');
+  removeGitConfigSection('git-mob');
 }
 
 function unsetCommitTemplate() {
-  exec('git config --unset commit.template');
+  removeGitConfigSection('commit');
 }
 
-function localGitConfigSectionEmpty(section) {
-  return exec(`git config --local --get-regexp '^${section}'`).status !== 0;
+function hasGitConfigSection(section) {
+  try {
+    const config = exec(`git config --get-regexp ${section}`);
+
+    return config.status !== 0 && config.stdout.trimEnd().length > 0;
+  } catch {
+    return false;
+  }
 }
 
 function safelyRemoveGitConfigSection(section) {
-  if (localGitConfigSectionEmpty(section)) {
-    exec(`git config --remove-section ${section}`);
+  if (hasGitConfigSection(section)) {
+    removeGitConfigSection(section);
   }
 }
 
 function removeGitConfigSection(section) {
-  exec(`git config --remove-section ${section}`);
+  exec(`git config --global --remove-section ${section}`);
 }
 
 function setGitMessageFile() {
@@ -108,7 +114,9 @@ function readCoauthorsFile() {
 
 function deleteCoauthorsFile() {
   try {
-    fs.unlinkSync(process.env.GITMOB_COAUTHORS_PATH);
+    if (fs.existsSync(process.env.GITMOB_COAUTHORS_PATH)) {
+      fs.unlinkSync(process.env.GITMOB_COAUTHORS_PATH);
+    }
   } catch (error) {
     console.warn('Failed to delete .git-coauthors file.', error.message);
   }
@@ -116,14 +124,45 @@ function deleteCoauthorsFile() {
 
 function deleteGitMessageFile() {
   try {
-    fs.unlinkSync(process.env.GITMOB_MESSAGE_PATH);
+    if (fs.existsSync(process.env.GITMOB_MESSAGE_PATH)) {
+      fs.unlinkSync(process.env.GITMOB_MESSAGE_PATH);
+    }
   } catch (error) {
     console.warn('Failed to delete .gitmessage file.', error.message);
   }
 }
 
 function exec(command) {
-  return spawnSync(command, { encoding: 'utf8', shell: true });
+  const spawnString = spawnSync(command, { encoding: 'utf8', shell: true });
+
+  if (spawnString.status !== 0) {
+    throw new Error(`GitMob handleResponse: "${command}" 
+    stdout: ${spawnString.stdout}
+    ---
+    stderr: ${spawnString.stderr}`);
+  }
+
+  return spawnString;
+}
+
+const testDir = './test-env';
+const repoDir = process.cwd();
+function setup() {
+  try {
+    if (!fs.existsSync(testDir)) {
+      fs.mkdirSync(testDir);
+    }
+  } catch (error) {
+    console.log(error);
+  }
+
+  process.chdir(testDir);
+  exec('git init -q');
+}
+
+function tearDown() {
+  process.chdir(repoDir);
+  fs.rmSync(testDir, { recursive: true });
 }
 
 module.exports = {
@@ -133,7 +172,6 @@ module.exports = {
   removeCoAuthors,
   unsetCommitTemplate,
   safelyRemoveGitConfigSection,
-  removeGitConfigSection,
   setGitMessageFile,
   readGitMessageFile,
   setCoauthorsFile,
@@ -141,5 +179,7 @@ module.exports = {
   deleteGitMessageFile,
   deleteCoauthorsFile,
   exec,
-  retainLocalAuthor
+  retainLocalAuthor,
+  setup,
+  tearDown
 };
